Crate bevy_rapier3d

Official integration of Rapier to the Bevy game engine

Rapier is a set of two Rust crates rapier2d and rapier3d for efficient cross-platform physics simulation. It target application include video games, animation, robotics, etc.

The bevy_rapier projects implements two other crates bevy_rapier2d and bevy_rapier3d which defines physics plugins for the Bevy game engine.

User documentation for bevy_rapier is on the official Rapier site.

Re-exports

pub extern crate nalgebra as na;

pub extern crate rapier3d as rapier;

pub use rapier::parry;

Modules

dynamics

Components related to physics dynamics (rigid-bodies, velocities, etc.)

geometry

Components related to physics geometry (colliders, collision-groups, etc.)

math

Type aliases to select the right vector/rotation types based on the dimension used by the engine.

pipeline

Components and resources related to the physics simulation workflow (events, hooks, etc.)

plugin

The physics plugin and systems.

prelude

Groups the most often used types.

utils

Miscellaneous helper functions.