Module bevy_rapier3d::dynamics

Components related to physics dynamics (rigid-bodies, velocities, etc.)

Structs

Ccd

Information used for Continuous-Collision-Detection.

Damping

Damping factors to gradually slow down a rigid-body.

Dominance

The dominance groups of a rigid-body.

ExternalForce

Constant external forces applied continuously to a rigid-body.

ExternalImpulse

Instantaneous external impulse applied continuously to a rigid-body.

FixedJoint

A fixed joint, locks all relative motion between two bodies.

FixedJointBuilder

Create fixed joints using the builder pattern.

GenericJoint

The description of any joint.

GenericJointBuilder

Create generic joints using the builder pattern.

GravityScale

Gravity is multiplied by this scaling factor before it’s applied to this rigid-body.

ImpulseJoint

An impulse-based joint attached to two entities.

LockedAxes

Flags affecting the behavior of the constraints solver for a given contact manifold.

MassProperties

Center-of-mass, mass, and angular inertia.

MultibodyJoint

An joint based on generalized coordinates, attached to two entities.

PrismaticJoint

A prismatic joint, locks all relative motion between two bodies except for translation along the joint’s principal axis.

PrismaticJointBuilder

Create prismatic joints using the builder pattern.

RapierImpulseJointHandle

The handle of an impulse joint added to the physics scene.

RapierMultibodyJointHandle

The handle of a multibody joint added to the physics scene.

RapierRigidBodyHandle

The Rapier handle of a rigid-body that was inserted to the physics scene.

ReadMassProperties

Center-of-mass, mass, and angular inertia.

RevoluteJoint

A revolute joint, locks all relative motion except for rotation along the joint’s principal axis.

RevoluteJointBuilder

Create revolute joints using the builder pattern.

Sleeping

The activation status of a body.

SphericalJoint

A spherical joint, locks all relative translations between two bodies.

SphericalJointBuilder

Create spherical joints using the builder pattern.

TransformInterpolation

If the TimestepMode::Interpolated mode is set and this component is present, the associated rigid-body will have its position automatically interpolated between the last two rigid-body positions set by the physics engine.

Velocity

The velocity of a rigid-body.

Enums

AdditionalMassProperties

Mass-properties of a rigid-body, added to the contributions of its attached colliders.

CoefficientCombineRule

Rules used to combine two coefficients.

RigidBody

A rigid-body.