Module bevy_rapier3d::geometry

Components related to physics geometry (colliders, collision-groups, etc.)

Modules

shape_views

Wrappers around Rapier shapes to access their properties.

Structs

ActiveCollisionTypes

Flags affecting whether or not collision-detection happens between two colliders depending on the type of rigid-bodies they are attached to.

ActiveEvents

Flags affecting the events generated for this collider.

ActiveHooks

Flags affecting the behavior of the constraints solver for a given contact manifold.

AsyncCollider

A component which will be replaced by the specified collider type after the referenced mesh become available.

AsyncSceneCollider

A component which will be replaced the specified collider types on children with meshes after the referenced scene become available.

Collider

A geometric entity that can be attached to a body so it can be affected by contacts and intersection queries.

CollidingEntities

Component which will be filled (if present) with a list of entities with which the current entity is currently in contact.

CollisionGroups

Pairwise collision filtering using bit masks.

ContactForceEventThreshold

The total force magnitude beyond which a contact force event can be emitted.

Friction

The friction affecting a collider.

InteractionGroups

Pairwise filtering using bit masks.

PointProjection

Result of the projection of a point on a shape.

RapierColliderHandle

The Rapier handle of a collider that was inserted to the physics scene.

RayIntersection

Structure containing the result of a successful ray cast.

Restitution

The restitution affecting a collider.

Sensor

Indicates whether or not the collider is a sensor.

SolverFlags

Flags affecting the behavior of the constraints solver for a given contact manifold.

SolverGroups

Pairwise constraints resolution filtering using bit masks.

Toi

The result of a time-of-impact (TOI) computation.

TriMeshFlags

The status of the cell of an heightfield.

VHACDParameters

Parameters controlling the VHACD convex decomposition.

Enums

ColliderMassProperties

Custom mass-properties of a collider.

ColliderScale

Overwrites the default application of [GlobalTransform::scale] to collider shapes.

ColliderView

Read-only access to the properties of a collider.

ComputedColliderShape

Shape type based on a Bevy mesh asset.

FillMode

Controls how the voxelization determines which voxel needs to be considered empty, and which ones will be considered full.

TOIStatus

The status of the time-of-impact computation algorithm.