Module bevy_rapier3d::geometry::shape_views

Wrappers around Rapier shapes to access their properties.

Structs

BallView

Read-only access to the properties of a ball.

BallViewMut

Read-write access to the properties of a ball.

CapsuleView

Read-only access to the properties of a capsule.

CapsuleViewMut

Read-write access to the properties of a capsule.

CompoundView

Read-only access to the properties of a compound shape.

ConeView

Read-only access to the properties of a cone.

ConeViewMut

Read-write access to the properties of a cone.

ConvexPolyhedronView

Read-only access to the properties of a convex polyhedron.

ConvexPolyhedronViewMut

Read-write access to the properties of a convex polyhedron.

CuboidView

Read-only access to the properties of a cuboid.

CuboidViewMut

Read-write access to the properties of a cuboid.

CylinderView

Read-only access to the properties of a cylinder.

CylinderViewMut

Read-write access to the properties of a cylinder.

HalfSpaceView

Read-only access to the properties of a half-space.

HalfSpaceViewMut

Read-write access to the properties of a half-space.

HeightFieldCellStatus

The status of the cell of an heightfield.

HeightFieldView

Read-only access to the properties of a heightfield.

HeightFieldViewMut

Read-write access to the properties of a heightfield.

PolylineView

Read-only access to the properties of a polyline.

PolylineViewMut

Read-write access to the properties of a polyline.

RoundConeView

Read-only access to the properties of a round shape.

RoundConeViewMut

Read-write access to the properties of a round shape.

RoundConvexPolyhedronView

Read-only access to the properties of a round shape.

RoundConvexPolyhedronViewMut

Read-write access to the properties of a round shape.

RoundCuboidView

Read-only access to the properties of a round shape.

RoundCuboidViewMut

Read-write access to the properties of a round shape.

RoundCylinderView

Read-only access to the properties of a round shape.

RoundCylinderViewMut

Read-write access to the properties of a round shape.

RoundTriangleView

Read-only access to the properties of a round shape.

RoundTriangleViewMut

Read-write access to the properties of a round shape.

SegmentView

Read-only access to the properties of a segment.

SegmentViewMut

Read-write access to the properties of a segment.

TriMeshView

Read-only access to the properties of a triangle mesh.

TriMeshViewMut

Read-write access to the properties of a triangle mesh.

TriangleView

Read-only access to the properties of a triangle.

TriangleViewMut

Read-write access to the properties of a triangle.

Enums

ColliderView

Read-only access to the properties of a collider.