Struct bevy_rapier3d::geometry::Collider

pub struct Collider {
    pub raw: SharedShape,
    /* private fields */
}

A geometric entity that can be attached to a body so it can be affected by contacts and intersection queries.

Fields
raw: SharedShape

The raw shape from Rapier.

Implementations

impl Collider

pub fn scale(&self) -> Vect

The scaling factor that was applied to this collider.

pub fn promote_scaled_shape(&mut self)

This replaces the unscaled version of this collider by its scaled version, and resets self.scale() to 1.0.

pub fn compound(shapes: Vec<(Vect, Rot, Collider)>) -> Self

Initialize a new collider with a compound shape.

pub fn ball(radius: Real) -> Self

Initialize a new collider with a ball shape defined by its radius.

pub fn halfspace(outward_normal: Vect) -> Option<Self>

Initialize a new collider build with a half-space shape defined by the outward normal of its planar boundary.

pub fn cylinder(half_height: Real, radius: Real) -> Self

Initialize a new collider with a cylindrical shape defined by its half-height (along along the y axis) and its radius.

pub fn round_cylinder(
half_height: Real,
radius: Real,
border_radius: Real
) -> Self

Initialize a new collider with a rounded cylindrical shape defined by its half-height (along along the y axis), its radius, and its roundedness (the radius of the sphere used for dilating the cylinder).

pub fn cone(half_height: Real, radius: Real) -> Self

Initialize a new collider with a cone shape defined by its half-height (along along the y axis) and its basis radius.

pub fn round_cone(half_height: Real, radius: Real, border_radius: Real) -> Self

Initialize a new collider with a rounded cone shape defined by its half-height (along along the y axis), its radius, and its roundedness (the radius of the sphere used for dilating the cylinder).

pub fn capsule(a: Vect, b: Vect, radius: Real) -> Self

Initialize a new collider with a capsule shape.

pub fn capsule_x(half_height: Real, radius: Real) -> Self

Initialize a new collider with a capsule shape aligned with the x axis.

pub fn capsule_y(half_height: Real, radius: Real) -> Self

Initialize a new collider with a capsule shape aligned with the y axis.

pub fn capsule_z(half_height: Real, radius: Real) -> Self

Initialize a new collider with a capsule shape aligned with the z axis.

pub fn cuboid(hx: Real, hy: Real, hz: Real) -> Self

Initialize a new collider with a cuboid shape defined by its half-extents.

pub fn round_cuboid(hx: Real, hy: Real, hz: Real, border_radius: Real) -> Self

Initialize a new collider with a round cuboid shape defined by its half-extents and border radius.

pub fn segment(a: Vect, b: Vect) -> Self

Initializes a collider with a segment shape.

pub fn triangle(a: Vect, b: Vect, c: Vect) -> Self

Initializes a collider with a triangle shape.

pub fn round_triangle(a: Vect, b: Vect, c: Vect, border_radius: Real) -> Self

Initializes a collider with a triangle shape with round corners.

pub fn polyline(vertices: Vec<Vect>, indices: Option<Vec<[u32; 2]>>) -> Self

Initializes a collider with a polyline shape defined by its vertex and index buffers.

pub fn trimesh(vertices: Vec<Vect>, indices: Vec<[u32; 3]>) -> Self

Initializes a collider with a triangle mesh shape defined by its vertex and index buffers.

pub fn from_bevy_mesh(
mesh: &Mesh,
collider_shape: &ComputedColliderShape
) -> Option<Self>

Initializes a collider with a Bevy Mesh.

Returns None if the index buffer or vertex buffer of the mesh are in an incompatible format.

pub fn convex_decomposition(vertices: &[Vect], indices: &[[u32; 3]]) -> Self

Initializes a collider with a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts.

pub fn round_convex_decomposition(
vertices: &[Vect],
indices: &[[u32; 3]],
border_radius: Real
) -> Self

Initializes a collider with a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts dilated with round corners.

pub fn convex_decomposition_with_params(
vertices: &[Vect],
indices: &[[u32; 3]],
params: &VHACDParameters
) -> Self

Initializes a collider with a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts.

pub fn round_convex_decomposition_with_params(
vertices: &[Vect],
indices: &[[u32; 3]],
params: &VHACDParameters,
border_radius: Real
) -> Self

Initializes a collider with a compound shape obtained from the decomposition of the given trimesh (in 3D) or polyline (in 2D) into convex parts dilated with round corners.

pub fn convex_hull(points: &[Vect]) -> Option<Self>

Initializes a new collider with a 2D convex polygon or 3D convex polyhedron obtained after computing the convex-hull of the given points.

pub fn round_convex_hull(points: &[Vect], border_radius: Real) -> Option<Self>

Initializes a new collider with a round 2D convex polygon or 3D convex polyhedron obtained after computing the convex-hull of the given points. The shape is dilated by a sphere of radius border_radius.

pub fn convex_mesh(points: Vec<Vect>, indices: &[[u32; 3]]) -> Option<Self>

Creates a new collider that is a convex polyhedron formed by the given triangle-mesh assumed to be convex (no convex-hull will be automatically computed).

pub fn round_convex_mesh(
points: Vec<Vect>,
indices: &[[u32; 3]],
border_radius: Real
) -> Option<Self>

Creates a new collider that is a round convex polyhedron formed by the given triangle-mesh assumed to be convex (no convex-hull will be automatically computed). The triangle mesh shape is dilated by a sphere of radius border_radius.

pub fn heightfield(
heights: Vec<Real>,
num_rows: usize,
num_cols: usize,
scale: Vector<Real>
) -> Self

Initializes a collider with a heightfield shape defined by its set of height (in column-major format) and a scale factor along each coordinate axis.

pub fn as_typed_shape(&self) -> ColliderView<'_>

Takes a strongly typed reference of this collider.

pub fn as_unscaled_typed_shape(&self) -> ColliderView<'_>

Takes a strongly typed reference of the unscaled version of this collider.

pub fn as_ball(&self) -> Option<BallView<'_>>

Downcast this collider to a ball, if it is one.

pub fn as_cuboid(&self) -> Option<CuboidView<'_>>

Downcast this collider to a cuboid, if it is one.

pub fn as_capsule(&self) -> Option<CapsuleView<'_>>

Downcast this collider to a capsule, if it is one.

pub fn as_segment(&self) -> Option<SegmentView<'_>>

Downcast this collider to a segment, if it is one.

pub fn as_triangle(&self) -> Option<TriangleView<'_>>

Downcast this collider to a triangle, if it is one.

pub fn as_trimesh(&self) -> Option<TriMeshView<'_>>

Downcast this collider to a triangle mesh, if it is one.

pub fn as_polyline(&self) -> Option<PolylineView<'_>>

Downcast this collider to a polyline, if it is one.

pub fn as_halfspace(&self) -> Option<HalfSpaceView<'_>>

Downcast this collider to a half-space, if it is one.

pub fn as_heightfield(&self) -> Option<HeightFieldView<'_>>

Downcast this collider to a heightfield, if it is one.

pub fn as_compound(&self) -> Option<CompoundView<'_>>

Downcast this collider to a compound shape, if it is one.

pub fn as_convex_polyhedron(&self) -> Option<ConvexPolyhedronView<'_>>

Downcast this collider to a convex polyhedron, if it is one.

pub fn as_cylinder(&self) -> Option<CylinderView<'_>>

Downcast this collider to a cylinder, if it is one.

pub fn as_cone(&self) -> Option<ConeView<'_>>

Downcast this collider to a cone, if it is one.

pub fn as_ball_mut(&mut self) -> Option<BallViewMut<'_>>

Downcast this collider to a mutable ball, if it is one.

pub fn as_cuboid_mut(&mut self) -> Option<CuboidViewMut<'_>>

Downcast this collider to a mutable cuboid, if it is one.

pub fn as_capsule_mut(&mut self) -> Option<CapsuleViewMut<'_>>

Downcast this collider to a mutable capsule, if it is one.

pub fn as_segment_mut(&mut self) -> Option<SegmentViewMut<'_>>

Downcast this collider to a mutable segment, if it is one.

pub fn as_triangle_mut(&mut self) -> Option<TriangleViewMut<'_>>

Downcast this collider to a mutable triangle, if it is one.

pub fn as_trimesh_mut(&mut self) -> Option<TriMeshViewMut<'_>>

Downcast this collider to a mutable triangle mesh, if it is one.

pub fn as_polyline_mut(&mut self) -> Option<PolylineViewMut<'_>>

Downcast this collider to a mutable polyline, if it is one.

pub fn as_halfspace_mut(&mut self) -> Option<HalfSpaceViewMut<'_>>

Downcast this collider to a mutable half-space, if it is one.

pub fn as_heightfield_mut(&mut self) -> Option<HeightFieldViewMut<'_>>

Downcast this collider to a mutable heightfield, if it is one.

pub fn as_cylinder_mut(&mut self) -> Option<CylinderViewMut<'_>>

Downcast this collider to a mutable cylinder, if it is one.

pub fn as_cone_mut(&mut self) -> Option<ConeViewMut<'_>>

Downcast this collider to a mutable cone, if it is one.

pub fn set_scale(&mut self, scale: Vect, num_subdivisions: u32)

Set the scaling factor of this shape.

If the scaling factor is non-uniform, and the scaled shape can’t be represented as a supported smooth shape (for example scalling a Ball with a non-uniform scale results in an ellipse which isn’t supported), the shape is approximated by a convex polygon/convex polyhedron using num_subdivisions subdivisions.

pub fn project_local_point_with_max_dist(
&self,
pt: Vect,
solid: bool,
max_dist: Real
) -> Option<PointProjection>

Projects a point on self, unless the projection lies further than the given max distance.

The point is assumed to be expressed in the local-space of self.

pub fn project_point_with_max_dist(
&self,
translation: Vect,
rotation: Rot,
pt: Vect,
solid: bool,
max_dist: Real
) -> Option<PointProjection>

Projects a point on self transformed by m, unless the projection lies further than the given max distance.

pub fn project_local_point(&self, pt: Vect, solid: bool) -> PointProjection

Projects a point on self.

The point is assumed to be expressed in the local-space of self.

pub fn project_local_point_and_get_feature(
&self,
pt: Vect
) -> (PointProjection, FeatureId)

Projects a point on the boundary of self and returns the id of the feature the point was projected on.

pub fn distance_to_local_point(&self, pt: Vect, solid: bool) -> Real

Computes the minimal distance between a point and self.

pub fn contains_local_point(&self, pt: Vect) -> bool

Tests if the given point is inside of self.

pub fn project_point(
&self,
translation: Vect,
rotation: Rot,
pt: Vect,
solid: bool
) -> PointProjection

Projects a point on self transformed by m.

pub fn distance_to_point(
&self,
translation: Vect,
rotation: Rot,
pt: Vect,
solid: bool
) -> Real

Computes the minimal distance between a point and self transformed by m.

pub fn project_point_and_get_feature(
&self,
translation: Vect,
rotation: Rot,
pt: Vect
) -> (PointProjection, FeatureId)

Projects a point on the boundary of self transformed by m and returns the id of the feature the point was projected on.

pub fn contains_point(&self, translation: Vect, rotation: Rot, pt: Vect) -> bool

Tests if the given point is inside of self transformed by m.

pub fn cast_local_ray(
&self,
ray_origin: Vect,
ray_dir: Vect,
max_toi: Real,
solid: bool
) -> Option<Real>

Computes the time of impact between this transform shape and a ray.

pub fn cast_local_ray_and_get_normal(
&self,
ray_origin: Vect,
ray_dir: Vect,
max_toi: Real,
solid: bool
) -> Option<RayIntersection>

Computes the time of impact, and normal between this transformed shape and a ray.

pub fn intersects_local_ray(
&self,
ray_origin: Vect,
ray_dir: Vect,
max_toi: Real
) -> bool

Tests whether a ray intersects this transformed shape.

pub fn cast_ray(
&self,
translation: Vect,
rotation: Rot,
ray_origin: Vect,
ray_dir: Vect,
max_toi: Real,
solid: bool
) -> Option<Real>

Computes the time of impact between this transform shape and a ray.

pub fn cast_ray_and_get_normal(
&self,
translation: Vect,
rotation: Rot,
ray_origin: Vect,
ray_dir: Vect,
max_toi: Real,
solid: bool
) -> Option<RayIntersection>

Computes the time of impact, and normal between this transformed shape and a ray.

pub fn intersects_ray(
&self,
translation: Vect,
rotation: Rot,
ray_origin: Vect,
ray_dir: Vect,
max_toi: Real
) -> bool

Tests whether a ray intersects this transformed shape.