Module bevy_rapier3d::plugin::systems

Systems responsible for interfacing our Bevy components with the Rapier physics engine.

Functions

apply_collider_user_changes

System responsible for applying changes the user made to a collider-related component.

apply_initial_rigid_body_impulses

This applies the initial impulse given to a rigid-body when it is created.

apply_joint_user_changes

System responsible for applying changes the user made to a joint component.

apply_rigid_body_user_changes

System responsible for applying changes the user made to a rigid-body-related component.

apply_scale

System responsible for applying [GlobalTransform::scale] and/or ColliderScale to colliders.

init_async_colliders

System responsible for creating Collider components from AsyncCollider components if the corresponding mesh has become available.

init_async_scene_colliders

System responsible for creating Collider components from AsyncSceneCollider components if the corresponding scene has become available.

init_colliders

System responsible for creating new Rapier colliders from the related bevy_rapier components.

init_joints

System responsible for creating new Rapier joints from the related bevy_rapier components.

init_rigid_bodies

System responsible for creating new Rapier rigid-bodies from the related bevy_rapier components.

step_simulation

System responsible for advancing the physics simulation, and updating the internal state for scene queries.

sync_removals

System responsible for removing from Rapier the rigid-bodies/colliders/joints which had their related bevy_rapier components removed by the user (through component removal or despawn).

update_colliding_entities

Adds entity to CollidingEntities on starting collision and removes from it when the collision ends.

writeback_rigid_bodies

System responsible for writing the result of the last simulation step into our bevy_rapier components and the GlobalTransform component.

Type Definitions

ColliderComponents

Components related to colliders.

RigidBodyComponents

Components related to rigid-bodies.

RigidBodyWritebackComponents

Components that will be updated after a physics step.